Source tree cleanup: move headers to src/, rename tests → test
- Move all 18 headers from include/moqx/ next to their .cpp in src/
- Delete include/ directory
- Rename tests/ → test/
- Extract test targets into test/CMakeLists.txt via add_subdirectory
- Update all include paths: strip moqx/ prefix, use quoted includes
- Update CMakeLists include directories from include → src

Addresses #78.
