+++ src/MoqxPicoRelayServer.h
@@ -3,24 +3,28 @@
 #include <memory>
 
 #include <folly/executors/IOThreadPoolExecutor.h>
-#include <moqx/MoqxRelayContext.h>
-#include <moqx/config/config.h>
-#include <moqx/stats/StatsRegistry.h>
-#include <moxygen/MoQServer.h>
+#include "MoqxRelayContext.h"
+#include "config/config.h"
+#include <moxygen/events/MoQExecutor.h>
+#include <moxygen/openmoq/transport/pico/MoQPicoQuicEventBaseServer.h>
+#include <proxygen/lib/http/webtransport/WebTransport.h>
 
 namespace openmoq::moqx {
 
-class MoqxRelayServer : public moxygen::MoQServer {
+// MoQ relay server backed by the picoquic QUIC stack.
+//
+// Picoquic lacks packet steering across threads, so even though the full
+// IOThreadPoolExecutor is passed in (for API symmetry with MoqxRelayServer),
+// the server always pins itself to evbKAs[0] internally.
+class MoqxPicoRelayServer : public moxygen::MoQPicoQuicEventBaseServer {
 public:
-  MoqxRelayServer(
+  MoqxPicoRelayServer(
       const config::ListenerConfig& listenerCfg,
       std::shared_ptr<MoqxRelayContext> context,
       std::shared_ptr<folly::IOThreadPoolExecutor> ioExecutor
   );
 
-  ~MoqxRelayServer() override;
-
-  void setStatsRegistry(std::shared_ptr<stats::StatsRegistry> registry);
+  ~MoqxPicoRelayServer() override;
 
   // Preferred entry point: binds the address from the stored ListenerConfig.
   void start();
@@ -47,7 +51,6 @@ protected:
 private:
   config::ListenerConfig listenerCfg_;
   std::shared_ptr<MoqxRelayContext> context_;
-  std::shared_ptr<folly::IOThreadPoolExecutor> ioExecutor_;
 };
 
 } // namespace openmoq::moqx
