+++ src/ServiceMatcher.cpp
@@ -1,4 +1,4 @@
-#include <moqx/ServiceMatcher.h>
+#include "ServiceMatcher.h"
 
 #include <algorithm>
