+++ src/config/loader.cpp
@@ -1,4 +1,4 @@
-#include "moqx/config/loader/loader.h"
+#include "config/loader/loader.h"
 
 #include <stdexcept>
