+++ src/config/loader/string_literal.h
