+++ src/stats/BoundedHistogram.h
