+++ test/MoqxRelayContextTest.cpp
@@ -1,4 +1,4 @@
-#include <moqx/MoqxRelayContext.h>
+#include "MoqxRelayContext.h"
 
 #include <folly/portability/GMock.h>
 #include <folly/portability/GTest.h>
