+++ test/MoqxRelayTest.cpp
@@ -6,7 +6,7 @@
  * Copyright (c) OpenMOQ contributors.
  */
 
-#include <moqx/MoqxRelay.h>
+#include "MoqxRelay.h"
 
 #include "TestUtils.h"
 #include <folly/coro/BlockingWait.h>
