+++ test/config/loader_test.cpp
@@ -1,4 +1,4 @@
-#include <moqx/config/loader/loader.h>
+#include "config/loader/loader.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
