+++ test/test_admin_info.sh
