+++ test/test_admin_metrics.sh
