+++ test/test_relay_chain.sh
